+++ examples/structural_mechanics/deforming_plate/deforming_plate_dataset.py
@@ -15,20 +15,14 @@
 # limitations under the License.
 
 
-import functools
 import json
 import os
 
 import numpy as np
 import torch
 
-try:
-    import tensorflow.compat.v1 as tf
-except ImportError:
-    raise ImportError(
-        "Mesh Graph Net Datapipe requires the Tensorflow library. Install the "
-        + "package at: https://www.tensorflow.org/install"
-    )
+from tfrecord.torch.dataset import TFRecordDataset
+
 
 try:
     import dgl
@@ -41,9 +35,6 @@ except ImportError:
 from torch.nn import functional as F
 
 from physicsnemo.datapipes.gnn.utils import load_json, save_json
-
-# Hide GPU from visible devices for TF
-tf.config.set_visible_devices([], "GPU")
 
 
 class DeformingPlateDataset(DGLDataset):
@@ -65,7 +56,7 @@ class DeformingPlateDataset(DGLDataset):
     num_samples : int, optional
         Number of samples, by default 1000
     num_steps : int, optional
-        Number of time steps in each sample, by default 600
+        Number of time steps in each sample, by default 400
     noise_std : float, optional
         The standard deviation of the noise added to the "train" split, by default 0.003
     force_reload : bool, optional
@@ -80,7 +71,7 @@ class DeformingPlateDataset(DGLDataset):
         data_dir=None,
         split="train",
         num_samples=1000,
-        num_steps=600,
+        num_steps=400,
         noise_std=0.003,
         force_reload=False,
         verbose=False,
@@ -99,13 +90,23 @@ class DeformingPlateDataset(DGLDataset):
 
         print(f"Preparing the {split} dataset...")
         # create the graphs with edge features
-        dataset_iterator = self._load_tf_data(self.data_dir, self.split)
+         # Build TFRecordDataset from .tfrecord file
+        tfrecord = os.path.join(data_dir, f"{split}.tfrecord")
+        index = None  # or path to .index if you generated it
+        # Define the schema per meta.json
+        meta = json.load(open(os.path.join(data_dir, "meta.json")))
+        description = {k: "byte" for k in meta["field_names"]}  # raw bytes
+        self.torch_ds = TFRecordDataset(
+            tfrecord, index, description,
+            transform=lambda rec: self._decode_record(rec, meta)
+        )
         self.graphs, self.cells, self.node_type = [], [], []
-        noise_mask, self.rollout_mask = [], []
+        noise_mask, self.moving_points_mask, self.object_points_mask, self.clamped_points_mask = [], [], [], []
         self.mesh_pos = []
-        for i in range(self.num_samples):
-            data_np = dataset_iterator.get_next()
-            data_np = {key: arr[:num_steps].numpy() for key, arr in data_np.items()}
+        for i, rec in enumerate(self.torch_ds):
+            if i >= num_samples:
+                break
+            data_np = {k: v[:num_steps] for k, v in rec.items()}
             src, dst = self.cell_to_adj(data_np["cells"][0])  # assuming stationary mesh
             graph = self.create_graph(src, dst, dtype=torch.int32)
             graph = self.add_edge_features(graph, data_np["mesh_pos"][0])
@@ -117,7 +118,10 @@ class DeformingPlateDataset(DGLDataset):
             if self.split != "train":
                 self.mesh_pos.append(torch.tensor(data_np["mesh_pos"][0]))
                 self.cells.append(data_np["cells"][0])
-                self.rollout_mask.append(self._get_rollout_mask(node_type))
+                moving_points_mask, object_points_mask, clamped_points_mask = self._get_rollout_mask(node_type)
+                self.moving_points_mask.append(moving_points_mask)        
+                self.object_points_mask.append(object_points_mask)
+                self.clamped_points_mask.append(clamped_points_mask)
 
         # compute or load edge data stats
         if self.split == "train":
@@ -134,21 +138,21 @@ class DeformingPlateDataset(DGLDataset):
             )
 
         # create the node features
-        dataset_iterator = self._load_tf_data(self.data_dir, self.split)
         self.node_features, self.node_targets = [], []
-        for i in range(self.num_samples):
-            data_np = dataset_iterator.get_next()
-            data_np = {key: arr[:num_steps].numpy() for key, arr in data_np.items()}
+        for i, rec in enumerate(self.torch_ds):
+            if i >= num_samples:
+                break
+            data_np = {k: v[:num_steps] for k, v in rec.items()}
             features, targets = {}, {}
-            features["world_pos"] = self._drop_last(data_np["world_pos"])
-            targets["world_pos"] = self._push_forward_diff(data_np["world_pos"])
-            targets["stress"] = self._push_forward(data_np["stress"])
+            features["world_pos"] = self._drop_last(data_np["world_pos"]) # Shape: (num_steps-1, num_nodes, num_features)
+            targets["velocity"] = self._push_forward_diff(data_np["world_pos"]) # Shape: (num_steps-1, num_nodes, num_features)
+            targets["stress"] = self._push_forward(data_np["stress"]) # Shape: (num_steps-1, num_nodes, num_features)
 
             # add noise
-            if split == "train":
-                features["world_pos"], targets["world_pos"] = self._add_noise(
+            if split == "train":  # TODO: noise has to be added at each iteration during training
+                features["world_pos"], targets["velocity"] = self._add_noise(
                     features["world_pos"],
-                    targets["world_pos"],
+                    targets["velocity"],
                     self.noise_std,
                     noise_mask[i],
                 )
@@ -163,15 +167,10 @@ class DeformingPlateDataset(DGLDataset):
 
         # normalize node features
         for i in range(num_samples):
-            self.node_features[i]["world_pos"] = self.normalize_node(
-                self.node_features[i]["world_pos"],
-                self.node_stats["world_pos_mean"],
-                self.node_stats["world_pos_std"],
-            )
-            self.node_targets[i]["world_pos"] = self.normalize_node(
-                self.node_targets[i]["world_pos"],
-                self.node_stats["world_pos_diff_mean"],
-                self.node_stats["world_pos_diff_std"],
+            self.node_targets[i]["velocity"] = self.normalize_node(
+                self.node_targets[i]["velocity"],
+                self.node_stats["velocity_mean"],
+                self.node_stats["velocity_std"],
             )
             self.node_targets[i]["stress"] = self.normalize_node(
                 self.node_targets[i]["stress"],
@@ -182,26 +181,28 @@ class DeformingPlateDataset(DGLDataset):
     def __getitem__(self, idx):
         gidx = idx // (self.num_steps - 1)  # graph index
         tidx = idx % (self.num_steps - 1)  # time step index
-        graph = self.graphs[gidx]
-        node_features = torch.cat(
-            (self.node_features[gidx]["world_pos"][tidx], self.node_type[gidx]), dim=-1
-        )
+        graph = self.graphs[gidx].clone()
+        node_features = self.node_type[gidx].float()
         node_targets = torch.cat(
             (
-                self.node_targets[gidx]["world_pos"][tidx],
+                self.node_targets[gidx]["velocity"][tidx],
                 self.node_targets[gidx]["stress"][tidx],
             ),
             dim=-1,
         )
         graph.ndata["x"] = node_features
         graph.ndata["y"] = node_targets
+        graph.ndata["world_pos"] = self.node_features[gidx]["world_pos"][tidx]
         if self.split == "train":
             return graph
         else:
             graph.ndata["mesh_pos"] = self.mesh_pos[gidx]
             cells = self.cells[gidx]
-            rollout_mask = self.rollout_mask[gidx]
-            return graph, cells, rollout_mask
+            moving_points_mask = self.moving_points_mask[gidx]
+            object_points_mask = self.object_points_mask[gidx]
+            clamped_points_mask = self.clamped_points_mask[gidx]
+
+            return graph, cells, moving_points_mask, object_points_mask, clamped_points_mask
 
     def __len__(self):
         return self.length
@@ -230,20 +231,18 @@ class DeformingPlateDataset(DGLDataset):
 
     def _get_node_stats(self):
         stats = {
-            "world_pos_mean": 0,
-            "world_pos_meansqr": 0,
-            "world_pos_diff_mean": 0,
-            "world_pos_diff_meansqr": 0,
+            "velocity_mean": 0,
+            "velocity_meansqr": 0,
             "stress_mean": 0,
             "stress_meansqr": 0,
         }
         for i in range(self.num_samples):
-            stats["world_pos_mean"] += (
-                torch.mean(self.node_features[i]["world_pos"], dim=(0, 1))
+            stats["velocity_mean"] += (
+                torch.mean(self.node_targets[i]["velocity"], dim=(0, 1))
                 / self.num_samples
             )
-            stats["world_pos_meansqr"] += (
-                torch.mean(torch.square(self.node_features[i]["world_pos"]), dim=(0, 1))
+            stats["velocity_meansqr"] += (
+                torch.mean(torch.square(self.node_targets[i]["velocity"]), dim=(0, 1))
                 / self.num_samples
             )
             stats["stress_mean"] += (
@@ -254,54 +253,18 @@ class DeformingPlateDataset(DGLDataset):
                 torch.mean(torch.square(self.node_targets[i]["stress"]), dim=(0, 1))
                 / self.num_samples
             )
-            stats["world_pos_diff_mean"] += (
-                torch.mean(
-                    self.node_targets[i]["world_pos"],
-                    dim=(0, 1),
-                )
-                / self.num_samples
-            )
-            stats["world_pos_diff_meansqr"] += (
-                torch.mean(
-                    torch.square(self.node_targets[i]["world_pos"]),
-                    dim=(0, 1),
-                )
-                / self.num_samples
-            )
-        stats["world_pos_std"] = torch.sqrt(
-            stats["world_pos_meansqr"] - torch.square(stats["world_pos_mean"])
+        stats["velocity_std"] = torch.sqrt(
+            stats["velocity_meansqr"] - torch.square(stats["velocity_mean"])
         )
         stats["stress_std"] = torch.sqrt(
             stats["stress_meansqr"] - torch.square(stats["stress_mean"])
         )
-        stats["world_pos_diff_std"] = torch.sqrt(
-            stats["world_pos_diff_meansqr"] - torch.square(stats["world_pos_diff_mean"])
-        )
-        stats.pop("world_pos_meansqr")
+        stats.pop("velocity_meansqr")
         stats.pop("stress_meansqr")
-        stats.pop("world_pos_diff_meansqr")
 
         # save to file
         save_json(stats, "node_stats.json")
         return stats
-
-    def _load_tf_data(self, path, split):
-        """
-        Utility for loading the .tfrecord dataset in DeepMind's MeshGraphNet repo:
-        https://github.com/deepmind/deepmind-research/tree/master/meshgraphnets
-        Follow the instructions provided in that repo to download the .tfrecord files.
-        """
-        dataset = self._load_dataset(path, split)
-        dataset_iterator = tf.data.make_one_shot_iterator(dataset)
-        return dataset_iterator
-
-    def _load_dataset(self, path, split):
-        with open(os.path.join(path, "meta.json"), "r") as fp:
-            meta = json.loads(fp.read())
-        dataset = tf.data.TFRecordDataset(os.path.join(path, split + ".tfrecord"))
-        return dataset.map(
-            functools.partial(self._parse_data, meta=meta), num_parallel_calls=8
-        ).prefetch(tf.data.AUTOTUNE)
 
     @staticmethod
     def cell_to_adj(cells):
@@ -320,6 +283,7 @@ class DeformingPlateDataset(DGLDataset):
         torch.int32 can handle graphs with up to 2**31-1 nodes or edges.
         """
         graph = dgl.to_bidirected(dgl.graph((src, dst), idtype=dtype))
+        graph = dgl.to_simple(graph)
         return graph
 
     @staticmethod
@@ -384,64 +348,30 @@ class DeformingPlateDataset(DGLDataset):
 
     @staticmethod
     def _get_rollout_mask(node_type):
-        mask = torch.logical_or(
-            torch.eq(node_type, torch.zeros_like(node_type)),
-            torch.eq(
-                node_type,
-                torch.zeros_like(node_type) + 1,
-            ),
-        )
-        return mask
+        moving_points_mask = torch.eq(node_type, torch.zeros_like(node_type))
+        object_points_mask = torch.eq(node_type, torch.zeros_like(node_type) + 1)
+        clamped_points_mask = torch.eq(node_type, torch.zeros_like(node_type) + 3)
+        return moving_points_mask, object_points_mask, clamped_points_mask
 
     @staticmethod
-    def _add_noise(features, targets, noise_std, noise_mask):  # disabling for now
-        # noise = torch.normal(mean=0, std=noise_std, size=features.size())
-        # noise_mask = noise_mask.expand(features.size()[0], -1, 2)
-        # noise = torch.where(noise_mask, noise, torch.zeros_like(noise))
-        # features += noise
-        # targets -= noise
+    def _add_noise(features, targets, noise_std, noise_mask): 
+        noise = torch.normal(mean=0, std=noise_std, size=features.size())
+        noise_mask = noise_mask.expand(features.size()[0], -1, 3)
+        noise = torch.where(noise_mask, noise, torch.zeros_like(noise))
+        features += noise
+        targets -= noise
         return features, targets
 
-    @staticmethod
-    def _parse_data(p, meta):
-        outvar = {}
-        feature_dict = {k: tf.io.VarLenFeature(tf.string) for k in meta["field_names"]}
-        features = tf.io.parse_single_example(p, feature_dict)
-        for k, v in meta["features"].items():
-            data = tf.reshape(
-                tf.io.decode_raw(features[k].values, getattr(tf, v["dtype"])),
-                v["shape"],
-            )
-            if v["type"] == "static":
-                data = tf.tile(data, [meta["trajectory_length"], 1, 1])
-            elif v["type"] == "dynamic_varlen":
-                row_len = tf.reshape(
-                    tf.io.decode_raw(features["length_" + k].values, tf.int32), [-1]
-                )
-                data = tf.RaggedTensor.from_row_lengths(data, row_lengths=row_len)
-            outvar[k] = data
-        return outvar
+    
+    def _decode_record(self, rec_bytes, meta):
+        out = {}
+        for k, v in rec_bytes.items():
+            dtype = meta["features"][k]["dtype"]
+            shape = meta["features"][k]["shape"]
+            arr = np.frombuffer(v, dtype=getattr(np, dtype))
+            arr = arr.reshape(shape)
+            if meta["features"][k]["type"] == "static":
+                arr = np.tile(arr, (meta["trajectory_length"], 1, 1))
+            out[k] = arr
+        return out
 
-
-if __name__ == "__main__":
-    # Minimal example usage
-    # Make sure to set the correct data_dir to your .tfrecord files
-    data_dir = "/code/mnabian/plate/physicsnemo/examples/solid_mechanics/deforming_plate/raw_dataset/deforming_plate/deforming_plate"  # <-- change this to your data path
-
-    # Instantiate the dataset (with small numbers for a quick test)
-    dataset = DeformingPlateDataset(
-        name="deforming_plate_example",
-        data_dir=data_dir,
-        split="train",
-        num_samples=10,
-        num_steps=10,
-        force_reload=True,
-        verbose=True,
-    )
-
-    # Fetch a sample
-    sample = dataset[0]
-    print("Sample graph:", sample)
-    print("Number of nodes:", sample.number_of_nodes())
-    print("Node features shape:", sample.ndata["x"].shape)
-    print("Node targets shape:", sample.ndata["y"].shape)