+++ physicsnemo/models/gnn_layers/mesh_node_block.py
@@ -21,7 +21,7 @@ import torch.nn as nn
 from torch import Tensor
 
 from .mesh_graph_mlp import MeshGraphMLP
-from .utils import GraphType, aggregate_and_concat
+from .utils import GraphType, agg_concat_hetero_dgl, aggregate_and_concat
 
 
 class MeshNodeBlock(nn.Module):
@@ -89,3 +89,74 @@ class MeshNodeBlock(nn.Module):
         # update node features + residual connection
         nfeat_new = self.node_mlp(cat_feat) + nfeat
         return efeat, nfeat_new
+
+
+class HybridMeshNodeBlock(nn.Module):
+    """Node block used e.g. in GraphCast or MeshGraphNet
+    operating on a latent space represented by a mesh.
+    Use for heterogeneous graphs.
+
+    Parameters
+    ----------
+    aggregation : str, optional
+        Aggregation method (sum, mean) , by default "sum"
+    input_dim_nodes : int, optional
+        Input dimensionality of the node features, by default 512
+    input_dim_edges : int, optional
+        Input dimensionality of the edge features, by default 512
+    output_dim : int, optional
+        Output dimensionality of the node features, by default 512
+    hidden_dim : int, optional
+        Number of neurons in each hidden layer, by default 512
+    hidden_layers : int, optional
+        Number of neurons in each hidden layer, by default 1
+    activation_fn : nn.Module, optional
+       Type of activation function, by default nn.SiLU()
+    norm_type : str, optional
+        Normalization type ["TELayerNorm", "LayerNorm"].
+        Use "TELayerNorm" for optimal performance. By default "LayerNorm".
+    recompute_activation : bool, optional
+        Flag for recomputing activation in backward to save memory, by default False.
+        Currently, only SiLU is supported.
+    """
+
+    def __init__(
+        self,
+        aggregation: str = "sum",
+        input_dim_nodes: int = 512,
+        input_dim_edges: int = 512,
+        output_dim: int = 512,
+        hidden_dim: int = 512,
+        hidden_layers: int = 1,
+        activation_fn: nn.Module = nn.SiLU(),
+        norm_type: str = "LayerNorm",
+        recompute_activation: bool = False,
+    ):
+        super().__init__()
+        self.aggregation = aggregation
+
+        self.node_mlp = MeshGraphMLP(
+            input_dim=input_dim_nodes + input_dim_edges,
+            output_dim=output_dim,
+            hidden_dim=hidden_dim,
+            hidden_layers=hidden_layers,
+            activation_fn=activation_fn,
+            norm_type=norm_type,
+            recompute_activation=recompute_activation,
+        )
+
+    @torch.jit.ignore()
+    def forward(
+        self,
+        mesh_efeat: Tensor,
+        world_efeat: Tensor,
+        nfeat: Tensor,
+        graph: GraphType,
+    ) -> Tuple[Tensor, Tensor]:
+        # update edge features
+        cat_feat = agg_concat_hetero_dgl(
+            mesh_efeat, world_efeat, nfeat, graph, self.aggregation
+        )
+        # update node features + residual connection
+        nfeat_new = self.node_mlp(cat_feat) + nfeat
+        return mesh_efeat, world_efeat, nfeat_new